requirements/test.txt
django-nose==1.3
git+http://github.com/arthuralvim/django-nose.git
nose==1.3.4
nose-cov==1.6
nose-progressive==1.5.1
xtraceback==0.4.0-rc1
rudolf2==0.3
pylint==1.2.1
clonedigger==1.1.0
autopep8==1.0.4
pep8ify==0.0.12
pep8==1.5.7